Few changes after rebasing onto main.

+++ network/views.py
@@ -1,13 +1,13 @@
 """ View for network. """
+
 from celery import shared_task
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.decorators import api_view
 from orca_nw_lib.device import get_device_details
 from orca_nw_lib.discovery import discover_device
+from orca_nw_lib.device_db import delete_device
 
-from log_manager.decorators import log_request, log_task
-from network.util import add_msg_to_list, get_failure_msg, get_success_msg, save_log
 from log_manager.decorators import log_request, log_task
 from network.util import add_msg_to_list, get_failure_msg, get_success_msg, save_log
 
@@ -28,20 +28,22 @@ def delete_db(request):
     Returns:
         Response: The HTTP response object with the result of the deletion operation.
     """
+    result = []
     if request.method == "DELETE":
         req_data_list = (
             request.data if isinstance(request.data, list) else [request.data]
         )
         task_id = save_log(req_data_list, request.method)
-        task = delete_db_task.apply_async(
+        delete_db_task.apply_async(
             (req_data_list, request.method, task_id), task_id=task_id
         )
-        return Response({"result": "Task Queued"}, status=status.HTTP_100_CONTINUE)
+        add_msg_to_list(result,get_success_msg(request.method))
+        return Response({"result": result}, status=status.HTTP_100_CONTINUE)
 
 
 @shared_task
 @log_task
-def delete_db_task(request):
+def delete_db_task(request, http_method, task_id):
     """
     A function that deletes the database.
 
@@ -51,31 +53,38 @@ def delete_db_task(request):
     Returns:
         Response: The HTTP response object with the result of the deletion operation.
     """
-    from orca_nw_lib.device_db import delete_device
-
     result = []
+    http_status: bool = True
+    if http_method == "DELETE":
+        try:
+            req_data_list = (
+                request.data if isinstance(request.data, list) else [request.data]
+            )
+            for req_data in req_data_list:
+                device_ip = req_data.get("mgt_ip", "")
 
-    try:
-        req_data_list = (
-            request.data if isinstance(request.data, list) else [request.data]
-        )
-        for req_data in req_data_list:
-            device_ip = req_data.get("mgt_ip", "")
+                del_res = delete_device(device_ip)
 
-            del_res = delete_device(device_ip)
+                if del_res:
+                    add_msg_to_list(result, get_success_msg(request))
+                else:
+                    add_msg_to_list(
+                        result,
+                        get_failure_msg(Exception("Failed to Delete"), http_method),
+                    )
+                    http_status = False
 
-            if del_res:
-                add_msg_to_list(result, get_success_msg(request))
-            else:
-                add_msg_to_list(
-                    result, get_failure_msg(Exception("Failed to Delete"), request)
-                )
-
-    except Exception as e:
-        return Response(
-            {"result": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR
-        )
-    return Response({"result": "Success"}, status=status.HTTP_200_OK)
+        except Exception as e:
+            add_msg_to_list(
+                result, get_failure_msg(e, http_method)
+            )
+            http_status = False
+    return Response(
+        {"result": result},
+        status=(
+            status.HTTP_200_OK if http_status else status.HTTP_500_INTERNAL_SERVER_ERROR
+        ),
+    )
 
 
 @api_view(
@@ -100,7 +109,7 @@ def discover(request):
         )
 
         task_id = save_log(req_data_list, request.method)
-        task = discover_task.apply_async(
+        discover_task.apply_async(
             (req_data_list, request.method, task_id), task_id=task_id
         )
         return Response({"result": "Task Queued"}, status=status.HTTP_100_CONTINUE)
